poc/chunking_benchmark_v2/pyproject.toml
[project]
name = "chunking-benchmark-v2"
version = "0.1.0"
description = "Advanced chunking benchmark with token-level metrics"
requires-python = ">=3.11"
dependencies = [
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "torch>=2.0.0",
    "tiktoken>=0.5.0",
    "scikit-learn>=1.3.0",
    "rank-bm25>=0.2.2",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = []

[tool.setuptools.packages.find]
where = ["."]
include = ["strategies*", "evaluation*", "retrieval*"]
exclude = ["corpus*", "results*"]
